server/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
duckduckgo-search>=4.1.1
transformers>=4.37.0
torch>=2.0.0
soundfile>=0.12.1
numpy>=1.24.0
geocoder==1.38.1
requests==2.31.0
beautifulsoup4>=4.12.0